Realizacion de mejoras

+++ src/main/kotlin/com/example/pruebadegradleparatesis/Repository/RolRepository.kt
@@ -3,5 +3,5 @@ package com.example.pruebadegradleparatesis.Repository
 import com.example.pruebadegradleparatesis.Model.Rol
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface RolRepository:JpaRepository<Rol,Int> {
+interface RolRepository:JpaRepository<Rol,Long> {
 }
