+++ src/main/kotlin/com/example/pruebadegradleparatesis/Repository/UsuarioRepository.kt
@@ -4,5 +4,5 @@ import com.example.pruebadegradleparatesis.Model.Usuario
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.*
 
-interface UsuarioRepository:JpaRepository<Usuario,Int> {
+interface UsuarioRepository:JpaRepository<Usuario,Long> {
 }
